[checker] Move checker code into its own package
