+++ samlang-core-checker/__tests/type-identifier-replacer.test.ts
