+++ samlang-core-checker/__tests/type-resolver.test.ts
