+++ samlang-core-checker/global-typing-context-builder.ts
