+++ samlang-core-checker/package.json
@@ -1,21 +1,18 @@
 {
-  "name": "@dev-sam/samlang-core",
+  "name": "samlang-core-checker",
   "version": "0.0.1",
   "license": "AGPLv3",
   "scripts": {
     "compile": "tsc --incremental --noEmit"
   },
-  "devDependencies": {
-    "@types/jest": "^26.0.10",
-    "antlr4ts-cli": "0.5.0-alpha.3",
-    "typescript": "4.0.2"
-  },
   "dependencies": {
-    "antlr4ts": "0.5.0-alpha.3",
     "samlang-core-ast": "workspace:0.0.1",
-    "samlang-core-checker": "workspace:0.0.1",
     "samlang-core-errors": "workspace:0.0.1",
     "samlang-core-parser": "workspace:0.0.1",
     "samlang-core-utils": "workspace:0.0.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
+    "typescript": "4.0.2"
   }
 }
