+++ samlang-core-checker/type-identifier-replacer.ts
