+++ workspaces.json
@@ -8,6 +8,7 @@
         "samlang-core-ast",
         "samlang-core-errors",
         "samlang-core-parser",
+        "samlang-core-checker",
         "@dev-sam/samlang-core",
         "@dev-sam/samlang-cli"
       ]
@@ -19,6 +20,7 @@
         "samlang-core-ast",
         "samlang-core-errors",
         "samlang-core-parser",
+        "samlang-core-checker",
         "@dev-sam/samlang-core"
       ]
     },
@@ -29,6 +31,7 @@
         "samlang-core-ast",
         "samlang-core-errors",
         "samlang-core-parser",
+        "samlang-core-checker",
         "@dev-sam/samlang-core",
         "@dev-sam/samlang-demo"
       ]
@@ -38,6 +41,16 @@
       "dependencyChain": [
         "samlang-core-utils",
         "samlang-core-ast"
+      ]
+    },
+    "samlang-core-checker": {
+      "workspaceLocation": "samlang-core-checker",
+      "dependencyChain": [
+        "samlang-core-utils",
+        "samlang-core-ast",
+        "samlang-core-errors",
+        "samlang-core-parser",
+        "samlang-core-checker"
       ]
     },
     "samlang-core-errors": {
@@ -69,6 +82,7 @@
     "samlang-core-ast",
     "samlang-core-errors",
     "samlang-core-parser",
+    "samlang-core-checker",
     "@dev-sam/samlang-core",
     "@dev-sam/samlang-cli",
     "@dev-sam/samlang-demo"
